refactored Text class into an abstraction class
- added OpenGLText class
- text drawing now done thru RenderCore::submit
- made Texture class memory-trackable
