+++ src/RocketEngine/RocketEngine.vcxproj.filters
@@ -135,6 +135,9 @@
     <ClInclude Include="render\platform\OpenGL\OpenGLTexture.h">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="render\platform\OpenGL\OpenGLText.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="core\EngineCore.cpp">
@@ -243,6 +246,9 @@
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="render\Mesh.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="render\platform\OpenGL\OpenGLText.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
   </ItemGroup>
